Fix typo: Change "ServerScriptStorage" to "ServerScriptService" (#1064)
## Changes

This pull request corrects a typo in the code where
"ServerScriptStorage" was mistakenly used instead of
"ServerScriptService".

Source: https://create.roblox.com/docs/chat/examples/group-chat-tags

## Checks

By submitting your pull request for review, you agree to the following:

- [X] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [X] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [X] To the best of my knowledge, all proposed changes are accurate.