Update StarterGui.yaml (#1080)
Add warning for MakeChatSystemMessage regarding TextChatService
incompatibility; Using MakeChatSystemMessage on modern TextChatService
does not work

## Changes

I added a warning to `MakeChatSystemMessage` method about
LegacyChatService requirement.
This clarifies that this method only works with the deprecated
LegacyChatService and points users to the TextChatService alternative
'DisplaySystemMessage'.

By submitting your pull request for review, you agree to the following:

- [✅] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [✅] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [✅] To the best of my knowledge, all proposed changes are accurate.

---------

Co-authored-by: IgnisRBX <43388550+IgnisRBX@users.noreply.github.com>