Update StandalonePluginScripts.yaml (#1076)
## Changes

<!-- Please summarize your changes. -->

Additional information has been added to document the instance's
inconsistent behavior. This info should be useful, as Studio does not
provide any additional info on this, nor is there any additional
information online.

<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->

Changes are based on the following comments and personal observations of
the instance's functionality throughout the years :


https://devforum.roblox.com/t/insert-object-menu-does-not-have-standalonepluginscripts/1161930/4?u=matkeftw
https://devforum.roblox.com/t/selection-doesnt-work-with-standalonepluginscripts/1204790/2?u=matkeftw

## Checks

By submitting your pull request for review, you agree to the following:

- [X] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [X] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [X] To the best of my knowledge, all proposed changes are accurate.

---------