docs: fix multiple typos (#1101)
## Changes

<!-- Please summarize your changes. -->

I have fixed multiple typos, such as changing `A set of instructions can
be e re-used in different parts of a script.` to `A set of instructions
that can be re-used in different parts of a script.` and changing
`.**30**` to `**0.300**`.

<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.