Patch 2 (#1103)
## Changes

<!-- Please summarize your changes. -->

Fixes typos on the documentation for CurveAnimation.
- Fixes "AnimationClipProviderProvider:RegisterAnimationClip" to
"\`AnimationClipProvider:RegisterAnimationClip()\`"
- Fixes "\`AnimationClipProvider:AnimationClipAsync\`" to
"\`AnimationClipProvider:GetAnimationClipAsync()\`"
- These add backquotes and parentheses to make these link to other parts
of the wiki correctly.

<!-- Please link to any applicable information (forum posts, bug
reports, etc.). -->

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.

---------

Co-authored-by: IgnisRBX <43388550+IgnisRBX@users.noreply.github.com>