AudioRecorder and Issue Request (#1104)
## Changes

In-reality I wanted to ask question. To find a valid way to do it, I've
updated ``AudioRecorder``, so that now I can use this text field here,
and type my question. **Because issue requests are disabled.**


Info is from
<https://create.roblox.com/docs/en-us/release-notes/release-notes-671>
and <https://devforum.roblox.com/t/release-notes-for-671> and own
testing in the current public version through Fast Flags.

* <https://devforum.roblox.com/t/release-notes-for-671/3634028/50>
* <https://devforum.roblox.com/t/release-notes-for-671/3634028/61>

## Checks

By submitting your pull request for review, you agree to the following:

- [x] This contribution was created in whole or in part by me, and I
have the right to submit it under the terms of this repository's open
source licenses.
- [x] I understand and agree that this contribution and a record of it
are public, maintained indefinitely, and may be redistributed under the
terms of this repository's open source licenses.
- [x] To the best of my knowledge, all proposed changes are accurate.


&nbsp;

# Issue

Alright. Now, to the issue.

In
https://create.roblox.com/docs/en-us/reference/engine/classes/ProximityPrompt
for instance, you can find a text that says

> The example below generates a custom

which is in a Code Sample, and that's currently not a thing

But what I am trying to say is, that there are pages, that aren't in
this Repository. And they can not be edited :(
There were moments where I wanted to edit something. I've probably
forgotten. But where do I go then? 🤔 Do I actually just go on the
Developer Forums with a Bug Report for everything?

Please fix.

---------

Co-authored-by: IgnisRBX <43388550+IgnisRBX@users.noreply.github.com>
Co-authored-by: rbxphogen <89100144+rbxphogen@users.noreply.github.com>